requirements.txt
openai
chainlit
python-dotenv